test(renderer-jsx): split markdown.test.tsx into per-component tests
Co-locate each component's tests next to its source — `Heading.test.tsx`
next to `Heading.tsx`, etc. — matching `File.test.tsx`/`Type.test.tsx`.
